Beginning Test Page
1. added:
- index.html: page for testing purposes (js needs DOM to add elements
to. In actual deployment the    js will need to be able to interact with
any html page)
- main.scss: styling of html, body, and container testing elements as
well as font and sizing of other elements.
- main.js: main object 'wolfman' is called in the html with
'wolfman.setUp()' to initialize and create d3 elements and
heights/positioning of elements based on document height. Adds events
listeners to window scroll to detect when viewer scroll past
visualization elements; functions are then called that make these
figures/visualizations react based on the users scroll actions.